test: print stderr and stdout on docker command failure

+++ tests/common/docker.rs
@@ -1,6 +1,5 @@
 use std::process::{Command, Output};
 
-
 pub const IMAGE_TAG: &str = "nvhosts-test";
 
 pub struct RunOutput {
@@ -8,6 +7,7 @@ pub struct RunOutput {
     pub container_id: String,
 }
 
+/// Builds and tags the docker image
 pub fn build(tag: &str) -> Output {
     let output = Command::new("docker")
         .arg("build")
@@ -22,18 +22,12 @@ pub fn build(tag: &str) -> Output {
         .expect("failed to execute process");
 
     if !output.status.success() {
-        eprintln!(
-            "docker build stdout: {:}",
-            String::from_utf8_lossy(&output.stdout)
-        );
-        eprintln!(
-            "docker build stdout: {:}",
-            String::from_utf8_lossy(&output.stderr)
-        );
+        debug(&output, "build");
     }
     output
 }
 
+/// Performs a docker run of the given tag
 pub fn run(tag: &str) -> RunOutput {
     let output = Command::new("docker")
         .arg("run")
@@ -48,14 +42,7 @@ pub fn run(tag: &str) -> RunOutput {
         .expect("failed to start container");
 
     if !output.status.success() {
-        eprintln!(
-            "docker build stdout: {:}",
-            String::from_utf8_lossy(&output.stdout)
-        );
-        eprintln!(
-            "docker build stdout: {:}",
-            String::from_utf8_lossy(&output.stderr)
-        );
+        debug(&output, "run");
     }
 
     RunOutput {
@@ -64,20 +51,36 @@ pub fn run(tag: &str) -> RunOutput {
     }
 }
 
+/// Stops and removes the container with th given id.
+///
+/// # Panics
+///
+/// Panics if either the docker stop or docker rm command fails.
 pub fn clean(id: &String) {
-    Command::new("sh")
+    let output = Command::new("sh")
         .arg("-c")
         .arg(format!("docker stop {}", id))
         .output()
         .expect("failed to stop container");
 
-    Command::new("sh")
+    if !output.status.success() {
+        debug(&output, "stop");
+        panic!("failed to stop docker container {}", id)
+    }
+
+    let output = Command::new("sh")
         .arg("-c")
         .arg(format!("docker rm {}", id))
         .output()
         .expect("failed to remove container");
+
+    if !output.status.success() {
+        debug(&output, "remove");
+        panic!("failed to remove docker container {}", id)
+    }
 }
 
+/// Prints container logs on stderr.
 pub fn logs(id: &String) {
     let output = Command::new("sh")
         .arg("-c")
@@ -92,4 +95,25 @@ pub fn logs(id: &String) {
         );
     }
     eprintln!("docker logs: {:}", String::from_utf8_lossy(&output.stdout));
+}
+
+/// Prints the stderr and stdout of a given [`Output`].
+fn debug(
+    Output {
+        status: _,
+        stdout,
+        stderr,
+    }: &Output,
+    stage: &str,
+) {
+    eprintln!(
+        "docker {} stdout: {:}",
+        stage,
+        String::from_utf8_lossy(stdout)
+    );
+    eprintln!(
+        "docker {} stderr: {:}",
+        stage,
+        String::from_utf8_lossy(stderr)
+    );
 }
